Fix get_symmetry method.
Former-commit-id: bbd7de198fb05a851d27e1158d3246359be051cd [formerly d41374a2b662e30d58a010346ac6fcc3c690cd5c]
Former-commit-id: ac4a4546515a67332716c5de3a462856badc3d0b

+++ pymatgen/symmetry/finder.py
@@ -246,8 +246,8 @@ class SymmetryFinder(object):
         # Get number of symmetry operations and allocate symmetry operations
         # multi = spg.multiplicity(cell, positions, numbers, symprec)
         multi = 48 * self._structure.num_sites
-        rotation = np.zeros((multi, 3, 3), dtype=int)
-        translation = np.zeros((multi, 3))
+        rotation = np.zeros((multi, 3, 3), dtype='intc')
+        translation = np.zeros((multi, 3), dtype='double')
 
         num_sym = spg.symmetry(rotation, translation,
                                self._transposed_latt.copy(),
